Generate Scoper for core Kubernetes types
To enable using nomos vet without an APIServer, we need an alternative
of hard-coded scopes for common Kubernetes types.

This CL adds CoreScoper, and a binary + make target that generates it
from the output of `kubectl api-resources`. Note the instructions in
cmd/gen-core-scoper/main.go that detail how to set this up.

Per Brian, I've started Makefile.gen that includes our two current code
generation targets. If we add more, we may want to make a target just
for generating code.

Note that we DO NOT want to automatically have CI generate
api_resources.txt, but have it done manually/reviewed. This is because
we want to control what cluster is being connected to in order to
generate this output. Otherwise the generated core_scoper.go may vary
wildly depending on who runs it.

Bug: 146063950
Change-Id: Ibe937d6c82b38f68e0eaf825710fffc589b75297