Allow all valid RBAC metadata.names
The Kubernetes APIServer makes an exception for all rbac types - it is
much more permissive about what their metadata.names may be. We applied
the full check that Kubernetes does for all other types to rbac.

This CL adds the exception for rbac types, replicating what the
APIServer does. This lets users put things like colon and capital
letters in things like Role names, as is allows in Kubernetes.

Bug: 145704739
Change-Id: I92c489a513258b1e45ede1b281315514b4ddb92a