Make the CRDRemovalValidator nonhierarchical
This has several advantages, most notably the code related to this check
is one-third the size it was before, and is consolidated into one file.
(Most of the added code in this CL is test code)

Related: RawParser did not properly do this check before, as it didn't
read the CRDs that were already synced from the CRDClusterConfig. Also,
Parser had no proper test of this. Tests have been added to verify that
this functionality works for them and won't regress.

Note that this changes the signature of Parser.Parse so it doesn't take
the full AllConfigs, but just the list of CRDs we are currently syncing
to the cluster. This is to enable testing, as writing test code to do
the byte-encoding of the CRDs is too cumbersome. Also, it advertises
exactly what we actually need for parsing to callers.

Change-Id: Idbce4f18cb3f656b188e372f9dfeaa2520859726