Allow disabling throwing errors on unknown types
This is required for serverless nomos vet, but if included in that CL
the CL is too big.

This just makes it so validation otherwise happens normally, but can
optionally be set to not throw errors specifically if it doesn't
recognize a type.

Add V(6) logging messages for when these errors are ignored.

Change-Id: I95bdee38938b701f7fbf4a0e6d52fa37eb1fb9a6