Implement flat NamespaceSelector resolution
Part of implementing non-hierarchical NamespaceSelectors is the logic
that actually resolves them.

This CL removes unnecessary logic from the hierarchical NS functions
in order to make them general enough for reuse.

The only real logic addition is validating namespace-selector references
in a flat repo.

This doesn't actually use the new logic in RawParser as that would make
the CL too big.

Change-Id: I5c79a58aa602c451f5cac25cdd42aa13ee9b5163