Copy abstract resources in flat repositories
go/unstructured-repo-clusterselectors

Implementing NamespaceSelectors in flat repos requires having some way
of duplicating resources that don't declare metadata.namespace.

The logic implemented in this CL assumes (as is verified elsewhere) that
in a flat repository, every namespace-scoped object declares either
metadata.namespace or the namespace-selector annotation.

Thus, the intent of this logic is to merely copy objects with the
namespace-selector annotation to all declared Namespaces. The actual
namespace-selector resolution logic is in a different CL.

Change-Id: I824d455ef30ea68c89bf89b8b4a0392ed636e816