Move k8s dependencies to 1.13.12
We previously upgraded to 1.16.3 with the intention of aligning with
google3, but I since realized that google3 is actually on 1.13.12. There
are some incompatibilities between controller-runtime 0.4.0 (which
google3 also uses) and k8s 1.13.12, which also would tend to drive one
to not use both, but google3 has already identified the patches that
make the two compatible.

Change-Id: Ic2b24eb92baac5a8022f830811bd61846b0d3c5f