Refactor the syncer's RestartableManager for shared use.
This is part of a multi-change effort to introduce a controller for
annotating Gatekeeper's custom resources. For part 1, we are refactoring
the RestartableManager so that it does not have any syncer-specific
logic exposed. This will allow it to be used by the Gatekeeper controller to restart its own manager and controllers for constraints as they get created/deleted.

Bug:147386001

Change-Id: I819f9b3b46ccd0745c22edb55131446a4d30dc3b