Split test ConfigManagements by channel
We now have multiple tests that change the operator configuration by
applying a ConfigManagement spec. For the -latest target, this resulted
in flipping between the dev and stable channels during the test, and not
returning to the stable channel correctly.

To fix this, mirror the existing configmanagement yamls into two
subdirectories within the manifests directory, stable and dev. Depending on the
channel the test suite is configured to run against (--stable flag passed
to setup.sh), set an environment variable MANIFEST_DIR to the
corresponding subdir. Tests that reference these yamls should use that
environment variable to locate their manifests.

A couple of small questions and how this current patch addresses them:

> What about syllogi?
Syllogi uses only one manifest. It is stable because we want to test our
stable branch in syllogi. The corresponding manifest is thus in stable
and not in dev. If we expanded the number of tests we run in syllogi
significantly, it might make sense to have a syllogi-specific subdir.

> What about manifests without channels?
We currently have one of these, tinyproxy. In this patch its just
mirrored into both for simplicity. If we have more of these we could
have a "util" or similar manifest directory.

Bug: 147318379
Change-Id: If815538446275f6ec4972a623d2c576707614784