Don't unnecessarily call the APIServer
We get resources from the APIServer sometimes when we don't actually
need them. The only case where we should actually try to talk to the
APIServer is when a repo defines a CR without a corresponding CRD.

This CL removes preemptively getting the CRDs synced to the cluster,
waiting to see if it is actully necessary. Thus, the interface for
Parser.Parse now takes a function that can return the synced CRDs, and
implementations only call that method if they actually need to.

The logic for constructing the Scoper / getting synced CRDs is factored
into its own file since each step deserves explanation. This means we no
longer have duplicate code for structured/unstructured parser.

Fix gen-core-scoper to actually defined scopes for Nomos types. Before
it only generated Kubernetes types despite advertising that it includes
Nomos types.

This makes obsolete a bunch of code in testing/fake.go since the
CoreScoper now does most everything testing.Scoper() did.

Change-Id: I69962eef536bcb4eab42dcb08eeea65f25eb877b