Make nomos hydrate optionally hermetic
Add --no-api-server-check flag to nomos hydrate.

Note that by default nomos vet/hydrate only talks to the API Server if
it actually needs to, so either can be used if the repositories
guarantee that there are CRDs for all non-default CRs.

If you don't like the phrasing for the flag description, please suggest
an exact alternative rather than giving a vague suggestion.

Where documentation related to this CL gets put, or how it is worded is
fully out of scope of this CL. Tech writers / repository authors are
responsible for the exact wording of that documentation and where it is
located, and will not be decided in this CL.

User docs:
https://docs.google.com/document/d/1s4Kt4Jpy_KDzkNnXkVVAKsl-5flC2gW6vP_C2Up_cT4/edit?usp=sharing

Bug: 138386428
Change-Id: I05660f9ec7bba805d7c39b2e87282e9c49c80fd9