Remove Makefile.installer
The oidc parts of the build are causing problems in prow (flakey) and
aren't working locally.  They seem to be dead code, so I'm deleting
them.

Also, the chat WEBHOOK in the bless-release command has a token in it,
which for some reason started engaging one of the commit filters
installed in git as part of GoB.  As we never seem to bless releases
anymore and I've never even seen this chat-hook work, I just removed it.

BUG: 150152814
Change-Id: I3f32c45045527d1c893c2440437c92d513c7b6a5