Minimal symink CL
This is an alternative for comparison to http://tg/684270

This CL is the smallest that I can come up with which correctly resolves
symlinks and has an end-to-end test that ensures symlinks are resolved
correctly.

It does NOT resolve issues related to:
- the cmpath library needing refactoring
- the controller logic not handling paths consistently
- how we handle empty repositories
- git-sync/importer concurrency and symlinks changing targets

E2E tests pass locally, and the added tests fails (as expected) without
the change.

Note the deletion of Root.Rel() since it is no longer necessary.

Change-Id: Ic9e5a0fda515526ba5d067e2e439729e8c267868