Don't restart submanager constantly
The current patching logic returns true on success.

The reconciler says that it returns "true" if the resource was updated.

The submanager is restarted if any CRDs were updated.

The flaw in this is that if CRDs have a no-op patch, the patch function
returns true, and the submanager restarts. This begins the cycle again.

This CL modifies the patching function to return "false" if there was
nothing to do, causing the submanager restarting logic to behave
more appropriately.

This does not resolve b/151856369, but is a major step in that
direction. There are other paths that still cause the submanager to
restart, but this gets rid of two of them - specifically, thinking that
CRDs have been updated when really the patches were NoOps.

Bug: 151856369
Change-Id: Id828383dccceeab406f35a94d0a65c40dc2dfbca