Don't try to remove creationTimestamps from CRDs
This just triggers the submanager to restart.

This CL treats patches that remove the creationTimestamp as a noop,
preventing signalling that CRDs have been modified.

While there may be a more elegant way of doing this, for now this works
and slows down the memory leak by a lot.

Reference b/152322972 on code changed to fix restarts - it needs unit
tests.

Bug: 151856369

Change-Id: I9196831cb4bf626967f80ea7e688619d9678213b