Don't delete existing Namespace Labels&Annotations
This makes the Syncer fight over Namespaces with things. We didn't
intend to do that. This also makes the Syncer preserve Labels and
Annotations that users specify manually, which is what we want. Before
we just clobbered them and removed them.

Also add unit tests for withNamespaceConfigMeta since this was the
source of the issue. The tests aren't exhaustive, but it is obviously
better to have the new tests than not to.

Bug: 151856871

Change-Id: I9747c343eaa3436e029a788ded0b79f24035e153