Implement fight detection logic
Caution: the specific math requires understanding solving first-order
ordinary differential equations. Contact willbeason@ for a lecture if
necessary.

We need a way to detect when the Syncer is fighting over resources with
some other process. We have arbitrarily decided that more than five
updates per minute is sufficient to warn the user that something may be
wrong.

This CL implements fightDetector. As with any frequency estimation
algorithm, it is not good at detecting frequencies exactly at the
threshold. Such boundary conditions are explicitly not tested as the
behavior there is unimportant.

Relevant performance characteristics noted in the comments.

Bug: 152337178

Change-Id: I8151db3ea6458829c8c2379331d5b9bf2eb67d64