Only log about fights once per minute
Otherwise we're just committing logspam when there is a problem. Once
per minute should keep the deulge sane when there is a problem.

If desired I can refactor to add tests for the logging logic. (I think
it'd be overkill, but if you add the comment I'll add the tests.)

Bug: 152972453
Change-Id: I5b5c33ecabf9f057a26acdc4ecf4ff8b13121298