Forbid managed resources in unmanaged Namespaces
Now that we don't destroy never-declared labels/annotations on
Namespaces, it is safe to declare Namespaces with no labels/annotations,
and manually modify them on the cluster.

This CL causes a nomos vet error when a Namespaces is unmanaged, but it
contains managed resources.

Unmanaged Namespaces containing unmanaged resources are fine.

This side-steps several issues caused by booting up new clusters with
unmanaged Namespaces containing managed resources (are we supposed to
create it?)

The error message is not intended as complete documentation for
unmanaging Namespaces. That is to be done in the linked user
documentation per b/152998030.

Bug: 152997454

Change-Id: I249ffd16b37d6da8727c90985320d46e08916efb