Reduce Syncer test reliance on Mocks
go/when-to-use-gomock

The Mocks in the Syncer tests make test code and failures essentially
unradable. The errors are difficult for seasoned veterans of our
codebase to decipher, let alone new team members.

The mock tests make changes to our code take hours longer than it needs
to, and make our code more brittle since there are many parts that we
simply refuse to change because rewriting the tests is so painful.

Also, google3 hugely frowns on gomock usage for the above reasons.

This CL is the first start of the effort to kill all mocks in our
codebase, replacing the Sync cache with a fake.

Change-Id: I274b23ce7a6d8ab4cf6f20eeedc6e62d8f8048e4