Respect v1.CustomResourceDefinition in the Syncer
This does not fully resolve b/153375960, but is targeted to resolve
b/153365093. Such a change is far out of scope of oncall fixing such a
systemic code issue.

Specifically, forbid the Cluster reconciler from syncing any resources
that are in apiextensions that are called CustomResourceDefinition. This
fixes two of the four failing tests in 1.16.

See the below doc for details:
https://docs.google.com/document/d/1yxvusZaW7I0uiJUK7D9eEBBoS4eIkaqJxRd1hNEJ01w/edit#heading=h.afb28dhbj1t6

This does not fully unit-test the code in scheme.go. That is out of
scope of this change. I have only written tests for the cases I've
changed the behavior for or am verifying. The behavior is covered in
other tests, but there are not tests targeted towards this logic.

Bug: 153365093
Change-Id: I05cb0e3c9b22d4ef226f620c73142ec6b2efc9fd