Don't overspecify e2e tests
This makes them flaky with version changes.

This does resolve b/153365093 as there are no more e2e tests that fail
with 1.16. Obviously it does not resolve the systemic issues surrounding
upgrading CRDs from v1beta1 to v1.

As before, my full notes on how I proved this works now in 1.16 are in
this doc:
https://docs.google.com/document/d/1yxvusZaW7I0uiJUK7D9eEBBoS4eIkaqJxRd1hNEJ01w/edit#heading=h.afb28dhbj1t6

Bug: 153365093

Change-Id: I1679445a425fbcf264562dd84887cae3ad6889f3