Remove DiscoveryInterface mock
Write a fake.DiscoveryClient initialied with the types it will serve,
but otherwise provides the DiscoveryClient methods the Syncer uses.

Limit the DiscoveryInterface -> ServerResourcer in the Syncer since it
only uses one method, and unnecessarily mocking out all of those other
methods is wasteful and makes code more confusing.

This all allowed for a few hundred lines of test code to just be
deleted since they were for working with this mock. All we were using it
for was to get the types available for the test - we don't need to
overspecify how the Syncer gets the information.

This code removes a now-unnecessary mockCache.List call. This was there
since the Deployment version listed was in corev1, when Deployments are
now in appsv1. The disparity caused the Syncer to try to list the
non-existent Deployment type in tests.

Bug: 152983731

Change-Id: I6141e66af3c31b94c2abbd5f4e56c758160d693c