Replace MockRecorder with a fake EventRecorder
The fake.EventRecorder is more permissive in what it checks, but this is
reasonable. We care that the event is about a particular resource, not
that the resource has a very particular state.

In testing, it is difficult to specify the full state of the resource
an Event is about as the resource may be in a strange
partial-modification state. But also, we don't actually care.

The fake.EventRecorder ensures each Event is about the specified
resource and has the stated event type and reason.

Bug: 152983731
Change-Id: I233d3c736b5410b922fbbfd376932a71f6eb0ec3