Replace mock client with fake
This is huge because so many tests are patterned to fuss about exactly
how mock clients are called. This CL drastically reduces how much tests
overspecify what they're testing.

This is the new paradigm:

1. The API Server is in some state at the beginning of the test.
2. The test makes some change to the API Server.
3. We validate that the state of the API Server is what we expect.

Our stand-in, of course, is the new fake.Client to represent the state
of the API Server.

We don't care how exactly the code makes this happen, just that it does.

The fake.Client takes a lot of shortcuts because ... it is a fake for
use in testing. I've tried to make it extensible so that we can add more
behavior if necessary (e.g. we want to List a new type in
non-Unstructured format).

This all did lead me to define core.ID - a term we've been toying with
for a while since so much of the test code uses it. There's several
places in production code that could be refactored to also use it, but
that is out of scope of this already-massive CL.

Note that I've gotten ride of StatusWriter. At the end of the day, it's
just a subset of Client. We don't care *how* the status field of an
object was updated, just that it *was* updated. Maintaining both is both
paradigm-mixing and requires keeping a lot of duplicate code around.

Notably, differ_test and sync/reconciler_test no longer use mocks!
There's only two more mocks to get rid of, and they will trouble us no
more.

Bug: 152983731

Change-Id: I9e95805d1331f5cfa9b61931a7e3afcaef6891bb