Significantly simplify Applier
We used a total of four clients to perform patches, when we really only
need one. This eliminates two of the clients - one which was there just
because the other wasn't usable without the other.

Now everything directly interacts with the generic client, which is just
a light wrapper around a client.Client.

This doesn't directly remove mocks, but it means much less logic has to
be faked since there are now two fewer types in the mix.

Bug: 152983731

Change-Id: Ic6bc7a0292df4514c306b80b29696a755719b272