Replace mock Applier with fake
Modify fake.Client to accept Scheme as a parameter. This ensures we're
holding and comparing types consistently, as the various reconcilers are
inconsistent with whether they're using Unstructured or actual structs.
The test author shouldn't have to care about this distinction.

Have fake.Client generate a fake.Applier that wraps it - analogous to
what we actually do. This way we only have to consider the state of one
thing at the end of the test to understand how the API Server was
modified.

As with other CLs in this sequence, the trend is to make tests
declarative about the state of the API Server at the end of the test
rather than monitoring every call to it.

Bug: 152983731

Change-Id: I647e5ac5b300156e9973b8092fc54b49f05d50bf