Replace mock Cache with fake.Client
This CL eliminates our last usages of gomock.

The mock cache and fake client are just different views of the same data.
If we want to explicitly test for how we handle differences in data between
the cache and the API Server then ... we should have tests for that,
which is out of scope of this CL.

GenericCache is much more complex than it has to be. I've modified it to
just use ListOptions like the rest of the k8s API so it doesn't have to
do the weird validation and give the obtuse "no index for Namespace"
error.

The only reason we had the GenericCache interface was because we had to
mock it. Since GenericCache is really just a wrapper around a
client.Reader, I've modified it to be so. We don't actually use any of
the Informer methods on the GenericCache, so this doesn't change
anything but what can be passed into the functions that create the
reconcilers.

This change lets us cut down on several things we had to do for test
that made them more complex just to satisfy gomock. We no longer need
our own comparison functions, or to write random labels/annotations to
keep it from panicking when it sees an empty map when it wanted a nil
map.

While this does slightly modify production code, all that is being done
is restricting the interface functions ask for, so it shouldn't actually
result in behavior differences.

Bug: 152983731

Change-Id: I3a9f702fe6c2662e5003e08247b6463eb65dcf7e