Un-export types
Unnecessarily exported types make it harder for code analyzers to detect
dead code, and make it harder for maintainers to know exactly what is
using the code they're modifying. In general, we want to export as
little as possible so developers have less to reason about.

This CL un-exports about a dozen types and deletes methods that are
subsequently obviously unused.

While this conflicts with "return structs, accept interfaces", in these
cases there is no value in returning the structs. The types export no
members and provide no methods in addition to the ones they define for
the interface they're implementing.

Change-Id: I5e6aa2a37b14de5ca151b0ea18c45cb59ce8c5ed