Clarify management annotation behavior.
No behavioral tests or refactoring, just adding comments to clarify why
tests seems to contradict each other. We're actually consistent, so I've
added comments that explicitly state how we are.

Essentially: If management is declared invalid, do nothing. This can
only happen if the ConfigManagment has been manipulated. If the state is
invalid on the cluster, either remove it if there is no declaration, or
overwrite it if the declared state is healthy.

Bug: 154535914
Change-Id: I7d5a6cdd75b58e0cfebcaed23b5f9103e886ecef