Un-export even more types
Third verse, same as the first.

Continue the quest to un-export unnecessarily-exported types and
methods. As usual, this resulted in discovering hundreds more lines of
code that are unnecessary.

Remove vt.Helper and just expose the methods directly. The struct itself
doesn't do anything but make it harder to generate test data.

Acknowledge that RestartChannel is really just an alias for a channel
with Restart() defined to make it satisfy an interface. We don't
otherwise maintain state in it, so it doesn't need to be anything more.

Deleted tests are so because they are the only consumers of their
otherwise-unused code.

Change-Id: I92302ea09dacf2db682bd721c87d685cbae30ced