Test metadata.labels for invalid values
The Kubernetes libraries incorrectly ignore non-string label values, so
we don't get an error from these libraries when such configuration is
ignored.

This CL explicitly checks for invalid label values like `true` instead
of `"true"` and suggests the appropriate fix.

This bug in the k8s libraries is fixed in 1.16, so we will be able to
remove this check eventually.

Bug: 154836948

Change-Id: I65c57e4e6c228e60e7309abaa81624e7983d9891