Prevent import that deletes all namespaces.
Another layer of defense against transient filesystem read errors during
import: we check to see if all NamespaceConfigs are being removed by an
import cycle. This is a dangerous operation so we ONLY allow it if the
cluster is going from one namespace to zero namespaces.

This one exception gives customers a way to wipe their cluster using
ConfigSync if that is actually what they intend to do. We also use this
mechanism to allow the e2e tests to wipe the cluster between tests to
decrease flakiness by reducing side effects.

Bug: 155392285
Bug: 154107250
Change-Id: I4ac4db9577b8e6484479e847b161612391eb9cc4