Fix: sync CRD gets stuck on a finalizer
The `sync` CRs have a `syncer` finalizer on them that isn't actually
removed when the syncer shuts down.  We thus loop through the `sync` CRs
and remove the finalizers from each via a `kubectl patch`.

This task is normally done imperatively by the operator as part of
uninstalling nomos.  See the `finalizeSyncs` function in
nomos-operator's `nomos_controller.go`.

Bug: 155073215
Change-Id: I370b3d0874f882e09481f5608e7fb2bfb4da54f7