Ensure all git files are imported
First off, use absolute paths with no symlinks whenever possible. This
eliminates whole classes of potential problems, as now we resolve
symlinks exactly once each pass.

The big change here is to the Parser, where we now require passing in
the repository root, and the list of files to read from. The
filesystem.Reader now requires listing each file to read from, and
throws an error if any file it was told to read from does not exist, or
it has another problem reading from it.

Vastly simplify the path library. All we care about is absolute paths,
except in the special case of ast.FileObjects, which use the relative
path to the policyDir root for bookkeeping.

Bug: 155385630
Change-Id: I72744f5352d7291f31dd1f45a160409cb758bbb1