Eliminate a race condition in the delete all namespaces test.
The safety namespace that is created in setup.bash to subvert our new
"never delete all namespaces on the cluster" was living longer than
expected, and frequently subverted the wrong test.  Adding a check on
namespace count and waiting for things to settle out seems to have made
this more deterministic.

Bug: 155512273
Change-Id: Idc52f9534892b6d30a50036e5be44b8558993eec