Make absolute/relative path types
This lets us join paths in a typesafe way, and makes it clear everywhere
whether we're dealing with a relative or absolute path.

This does introduce duplication between the Relative and Absolute types.
Defining an underlying type to reduce duplication doesn't reduce code
length, and only increases complexity. Most of these methods are very
small, so the danger is low.

This CL actually also fixes a few small bugs when running the CLI on
Windows - certain OS-specific paths were read as slash-delimited. Those
are now gone, and were easier to catch because they are explicit.

Bug: 155510326
Change-Id: I0efb636b42f299dc91639d6f6d29dcac0800b903