Move vet and init CLI tests to Go
These tests take a lot of time to run as E2E tests in bats, are brittle,
and difficult to understand.

This CL transforms the tests into Go tests. The only logic not tested by
these tests is internal Cobra logic. Otherwise, this fully captures the
flag parsing and filesystem operations used in the e2e tests. If we want
tests that run against the compiled CLI executable, we can do that
instead. (If so, I'll add a TODO to convert these tests in the next CL.)

Also, in doing so, I've simplified several of our tests that set up
filesystems in a specific way. We actually had three different
implementations of this, so I've pared it down to one.

The resulting Go tests execute in under a second, over two orders of
magnitude faster than the bats tests.

Part of this refactoring required switching vet/init to RunE, and
actually return the errors they produce rather than printing err and
dying. This is a much healthier pattern.

Change-Id: Ie5dde7456a05f7a22164b6feb53d57c7c01d2290