Support RAW_NOMOS in per_cluster_addressing.bats
This test does patch commands on the configManagement CR, so I replaced
that with patch commands on the configMap.  I thought it better to break
our existing pattern of applying configMap files so we didn't end up
with a bunch of configMap files that are only used in one test.

Bug: 155073215
Change-Id: I4761ba4b3df525695d29bb5e997725971559d9d0