No longer setting random permissions on created pre-release files.
We are removing the whitelist from the ACM GCS bucket, replacing it with
allAuthenticatedUsers:legacyObjectReader access.  For now this will ONLY
apply to the gs://config-management-release/released directory in GCS,
as we don't really want all of our build artifacts visible to anyone
with a gmail account, but it means we can stop setting the slew of
permissions on build artifacts.

Bug: 151333330
Change-Id: I759ddbbc3be6c555ef8a75047ca75677a6fb7832