Stop bugreport from failing if CM resources are not found on cluster.
Previously bugreport fails if the ConfigManagement object is not
found on the cluster. Additionally, each resource in FetchCMResources
(had it gotten that far) would add a not found error to the errorList
which triggers a message that the bugreport may have failed and tells
the user to try to find a partially generated file.

I am refactoring to treat these not founds as not errors as far as
bugreport is concerned since the bugreport is a diagnostic tool and it
is helpful to be able to reveal that required resources are unable
to be found.

TESTED=Ran bugreport locally and e2e
BUG=156016424

Signed-off-by: Craig Tabita <ctab@google.com>
Change-Id: I84acd5f73dc323b181f148f6a02baebe7c73109c