Begin cleanroom test
Write a test that:
1) Connects to an existing cluster
2) Ensures namespace "foo" does not exist
3) Creates namespace "foo"
4) Ensures namespace "foo" exists

In the future ACM will create the Namespace. This is a work-in-progress
proof-of-concept.

As-is, the test takes about 0.5s to complete but will probably take
~1.5min once all the TODOs are done.

Test does not run by default. Pass -e2e to run the test.

Attempting to write tests that use the library calls t.Fatal with a
message to add the flag. The safeguard is necessary to prevent test
authors from accidentally writing tests that don't use the e2e flag.

Bug: 156643819
Change-Id: Ibb12f2a253d407b9c06b99c650c1550fc732571c