Only check licenses once
We currently check the licenses of all of our dependencies about four
times. This is because we individually evaluate our binaries, rather
than just the list of our dependencies.

This CL uses go-licesnse on the output of `go list all`, greattly
speeding up this process. Rather than ~3min/binary, the whole process
takes about 3 minutes total.

This is technically unnecessarily checking some dependencies, but it's
faster and we don't currently have any non-binary dependencies that are
inviolation of OSPO requirements.

Change-Id: Id0318a124c3affafa64e84bfb541ce2c56f06620