Breaking tests out into functions.
The way some of these tests were running was only ever evaluating the
kubectl call once, and new things in the environment (in this case HNC)
was causing the test to fail.  Rewriting these as functions fixes the
issue.

This change also adds some timing to wait.bash (this was borrowed from
the operator).

Bug: 158000903
Change-Id: Ia74c8e091d7890c3ce8f0c171b28591eb21572fc