Install git-server in cleanroom test
Put git-server image in gcr.io/nomos-public. We're just using a
Dockerfile from a public git repo, so there isn't any reason to keep
this private. This also makes testing much easier and less flaky since
we don't have to worry about whether we have permission to access an
image of fully-open-source software.

Set up the git-server in the cleanroom test, and ensure it comes up
before proceeding with the test.

Change-Id: I38beef0276444638c0b47e794b411bd19f72045a