Change variable make targets to be explicit
We only ever use these for one thing and don't have plans to expand
them, so as-is just makes it harder to discover how things are done.

This CL just fills in the variables with the only values they're ever
used as.

This has the additional advantage of making those targets show up in
shell completion.

Change-Id: I0b7345cb7628efa7e5804cac86e5ea900ff41ada