Remove `token` related things from nomos
As part of decoupling nomos and the operator, we must move parts of the
nomos manifest that are not part of nomos' default, working state into
the operator.  This includes the support for passing in git-sync's
secrets as environment values in support of authentication strategy
`token`.

This will be accompanied by a CL in the operator, and the two CLs should
be merged together.

Bug: 158240726
Change-Id: I6667f508d1592a230c792d2aa3fb0c79c519f4f9