Finish Proof-of-Concept go e2e test
The go e2e test now actually tests Nomos behavior on a cluster.

Add instructions for how to run the test. These will probably change as
we develop this. For example, some of the manual steps should be make
targets.

Add debug flag that leaves clusters running and doesn't clean up
temporary directory. For now it isn't a problem, but eventually we may
want to make sure people don't leave 80 kind clusters running since
that'll be annoying. This mode halts test execution on the first
t.Fatal, and prints out how to connect to the cluster & etc.

Fix SSH. It turns out kubectl create secret is annoying to emulate, and
it doesn't expose the inner logic to outside consumers. So instead of
trying to do it ourselves, we're shelling out to kubectl to ensure we
create a valid set of secrets.

Write git.go that connects to the sot.git Git repository. I almost
succeeded in making this able to arbitrarily use repositories, but it
did something weird with git-server file permissions. It isn't worth
fixing now, but it's close to supporting multi-repo. We shell out for
git commands as the git libraries are difficult to configure ssh for,
and git-server required ssh authentication.

The SSH keys have to exist before the pods which consume them, as the
Pods don't always restart if the secrets they require don't exist yet.
Thus, we create the Namespaces first, then the SSH keys, and then
install git-server and ConfigSync.

Change-Id: Id9090e9b34b9d2e0793a43d33ee0993f783bb549