Remove operator support from e2e/
With the `raw` path now the sole testing path, it's time to rip out what
remains of the operator in e2e/.  This includes ripping out the
RAW_NOMOS flag, as that path is now the default.  I've left the target
names the same for now (i.e. including "raw" in the name), because I've
got a separate bug (b/158336498) for improving the target names.

Bug: 158246710
Change-Id: Idbb8d9a58a3209f1f3af87d04332a6927c88f664