Allow waiting for ConfigSync to Sync
Add .WaitForSync() that blocks until the Repo object reports that
ConfigSync has successfully synced everything declared in the repo.

Eventually we'll want things like .WaitForImport(), but those can easily
be done later.

Also add .RenewClient() that creates a Client with a new RESTMapper. The
RESTMapper isn't automatically updated when the cluster's API resources
change, so we have to create a new Client, which automatically has all
types available to the cluster when the Client is created. Test authors
only have to call RenewClient() directly if they're either applying a
CRD themselves or they expect another resource to create a CRD.

Change-Id: I55e55a3ecd6b954a8185c547b6b012f385619a2f