Make target to build & publish nomos-manifest.yaml
This new target (`make nomos-manifest`) isn't hooked up to anything, but
is the first ingredient in making a nomos manifest the way that nomos is
incorporated into the operator.  Further work will include building this
into the release postsubmit and building out a manual process for
associating these builds with a tag and proper version.

Bug: 158241701

Change-Id: Ia9892d12df169099e150f3c73a350903eda669e7