Reduce test logspam and make retrying return time
For client operations, only log create/update/delete and NOT get, as
this clogs up the logs when in a retry loop.

Return the duration that a command was retried on success to aid in
logging (and eventually tracking metrics).

Increase time to wait for sync to 2 minutes as 1 minute was not long
enough for acme tests.

Change-Id: I0e3a66f5ae8156d9cd1d5b70373503a7223b0074