First pass of RemediatorReconciler
Design: https://docs.google.com/document/d/1ncJqp-oKQpQgZHcex0YdKEYLIHAKsAcNLyrVwX2KA9Y/edit#heading=h.3m2y97jd932v

Write core of the new reconciler. Given an object to update, makes the
state of the object on the cluster match the declared state.

Change interface of DeclaredResources to use
core.Object/unstructured.Unstructured to make a lot of casting between
types unnecessary. All we really care about at that point is the
core.Object - we no longer need the path in the repository.

Reuse the syncer.Applier and the differ.Diff for calculating what type
of operation to perform, and how to perform it. All that's novel here is
that one method handles all types, and we don't use intermediate encoded
forms of objects, or NamespaceConfigs/ClusterConfigs.

Add Is() bool to the Error interface to make all KNV errors comparable
with errors.Is. This is the new go113+ way, and is just better in
general for testing.

This doesn't cover various edge cases such as special Namespaces. That
will be done in a later CL.

Bug: 157587458
Change-Id: I27fe9d99c43d18282752939f5e3fbc405a13c1c6