Remove an obsolite test long running test that is no longer needed.
This testcase was added in 78bfcf180b11bf8993c63d70a5af924ed09fba76 in response
to b/122529010 (which was about e2e tests not cleaning up properly b/c operator
issues) and was used for testing the operator to make sure it could handle
changing the cluster name over and over to ensure the operator would not lock
up.  This test typically takes around 5 minutes, so removing it will help us a
lot.

Change-Id: Ia31057f603b5d50d7758c834015b969a8261f65f