Use our own annotation for tracking "last applied".
Instead of using the same annotation as kubectl, we will store the most
recently applied configuration in our own annotation. This avoids
conflicts with other controllers that might use kubectl libraries (such
as the Config Sync operator).

This change includes a small bit of code for migration: if a managed
resource has the old kubectl annotation, then we move the value into the
new annotation.

Bug: 159164014
Change-Id: I8faf8d8883b1ad842d55f432482c4048203b8760