Integrate remediator with Importer.
Rename "Remediator" to "reconciler" per offline discussion.

Remember: If you don't like the name of something, suggest an
alternative name instead of just commenting "this needs a better name".

Add hook for remediator into Importer. All it needs to be able to do is
update the declared resources. Add future hook for WatchManager. It
definitely needs to be able to write to the work queue via add, get
the list of GVKs, and know when the declarations have been updated. That
can all be done once WatchManager actually exists.

This is just hooks into the no-op Remediator, so there's no behavior to
test.

Add NoOp remediator for use when Remediator behavior is disabled.

Mark TODO for launching goroutines to reconcile objects.

Mark TODO for hook into WatchManager.

Gerrit ignores the git file move directives, so the diff here sucks
since remediator.go got renamed to reconciler.go, and there is now new
code in remediator.go.

Change-Id: If435b4e573d386d044c790eae131c3362007b802