Have bats testlib skip teardown if test is skipped.
This inverts the control for the test such that the setup / teardown hooks are
in lib/setup.bash which lets us skip both setup and teardown if the test is
skipped.  Prior to this, we were running teardown on any skipped test which
could take up to 14 seconds per testcase.  This adds up to 140 seconds for a
file with 10 testcases the user does not want to run which results in a
significant time penalty.

Change-Id: Id60ad91f380e9b98fd937dbdaa48effcef12c473