Install nomos CLI locally
For the regular test target that rebuilds the nomos image, also install
the Nomos CLI on the machine running the test. This makes it available
for the cli.bats test.

Before it just complained that the CLI didn't exist, since the bats
tests aren't really running in the same environment as the one executing
the tests.

Change-Id: I872829fcb9b2569156d61d247c989f52a17e00c3